moved order-guarantee solution to Lesson 2 and relegated non-order-guarantee solution to Lesson 6.  Moved fanout instructions to teacher's guide.
