+++ Lesson2_CreateViewWithEventConsumer/winner-view/winner.js
@@ -12,17 +12,15 @@ const productPurchaseSchema = require('./product-purchase-schema.json')
 const constants = {
   // self
   MODULE: 'winner-view/winner.js',
+  UNKNOWN: 'UNKNOWN',
   NONE: 'NONE',
   // methods
   METHOD_REGISTER_CONTRIBUTOR: 'registerContributor',
   METHOD_UPDATE_SCORES_TABLES: 'updateScoresTable',
-  METHOD_GET_EVENTS_THEN_CREDIT: 'getEventsThenCredit',
-  METHOD_CREDIT_CONTRIBUTIONS: 'creditContributions',
   METHOD_UPDATE_PURCHASE_EVENT: 'updatePurchaseEvent',
   // resources
   TABLE_CONTRIBUTIONS_NAME: process.env.TABLE_CONTRIBUTIONS_NAME,
   TABLE_SCORES_NAME: process.env.TABLE_SCORES_NAME,
-  TABLE_EVENTS_NAME: process.env.TABLE_EVENTS_NAME,
 }
 
 const kh = new KH.KinesisSynchronousHandler(eventSchema, constants.MODULE)
@@ -57,23 +55,27 @@ const impl = {
 
     const updateCallback = (err) => {
       if (err) {
-        complete(`${constants.METHOD_REGISTER_CONTRIBUTOR} - errors updating DynamoDb: ${err}`)
+        if (err.code && err.code === 'ConditionalCheckFailedException') {
+          console.log(`${constants.METHOD_REGISTER_CONTRIBUTOR} - registration event has already been processed for ${role}.  Skipping.`)
+          complete()
+        } else {
+          complete(`${constants.METHOD_REGISTER_CONTRIBUTOR} - errors updating DynamoDb: ${err}`)
+        }
       } else {
-        const roleInfo = {}
-        roleInfo[role] = event.origin
-        impl.getEventsThenCredit(event.data.id, event.eventId, event.origin, roleInfo, complete)
+        complete()
       }
     }
 
-    // Record product's contributor registration
+    // Note if someone manually re-registers as the contributor later, it gets ignored and credit continues to accumulate to the original contributor (until we explicitly support re-shoots as a feature).
     const expression = [
       'set',
       '#c=if_not_exists(#c,:c),',
       '#cb=if_not_exists(#cb,:cb),',
       '#u=:u,',
       '#ub=:ub,',
-      '#ro=if_not_exists(#ro,:ro),',
-      '#ev=if_not_exists(#ev,:ev)',
+      '#ro=if_not_exists(#ro,:ro),', // We shouldn't have to block overwrites, because order guarantee should mean registration event will always come first and nature of hello-retail should not have any way to change a creator or photographer, once registered, BUT just saw this happen with photographers because someone went in and reset by hand.  Until we have re-shoots be a feature, all credit still goes to original contributor.
+      '#sc=if_not_exists(#sc,:zero),', // Score initializes at 0.  If this was an attempt to put in new contributor, then that gets blocked (no overwrites of contributors), so score should not reset to 0 (unless we later make re-shoots a feature).
+      '#ev=:ev',
     ]
     const attNames = {
       '#c': 'created',
@@ -81,7 +83,9 @@ const impl = {
       '#u': 'updated',
       '#ub': 'updatedBy',
       '#ro': role,
-      '#ev': `${role}EventId`,
+      '#sc': `${role}Score`, // These scores may be different for the two roles because some purchases may happen between the time that the creator and photographer registered.  Each score should be initialized only its registration event.
+      '#ev': 'lastEventId',
+      '#pe': 'photographerExists',
     }
     const attValues = {
       ':c': updated,
@@ -90,6 +94,21 @@ const impl = {
       ':ub': event.origin,
       ':ro': event.origin,
       ':ev': event.eventId,
+      ':zero': 0,
+    }
+
+    if (role === 'creator') {
+      attNames['#sp'] = 'photographerScore'
+      expression.push(', #sp=if_not_exists(#sp,:zero)') // To keep if_not_exists from blowing up, this must be defined before any purchase events happen
+      expression.push(', #pe=:pe')
+      attValues[':pe'] = -1
+    } else if (role === 'photographer') { // NB if there was no creation event but the photographer event made it on, the photographer should still be UNKNOWN, because the score has been accumulating as if he were UNKNOWN, but there is no easy way to overwrite the name with UNKNOWN if one pass, so will handle this when updatingScores instead.  Basically, if you see UNKNOWN in the creator field, ignore the person who is in the photographer field, if any.
+      attNames['#scr'] = 'creatorScore'
+      attNames['#cr'] = 'creator'
+      attValues[':unk'] = constants.UNKNOWN
+      expression.push(', #scr=if_not_exists(#scr,:zero)')
+      expression.push(', #cr=if_not_exists(#cr,:unk)') // To block any creator attribution for events without a creation event (i.e., log incomplete from the front for product)
+      expression.push('remove #pe') // REMOVE: If the attributes do not exist, nothing happens
     }
 
     const dbParamsContributions = {
@@ -98,6 +117,7 @@ const impl = {
         productId: event.data.id,
       },
       UpdateExpression: expression.join(' '),
+      ConditionExpression: 'attribute_not_exists(#ev) OR #ev < :ev',
       ExpressionAttributeNames: attNames,
       ExpressionAttributeValues: attValues,
       ReturnValues: constants.NONE,
@@ -105,256 +125,6 @@ const impl = {
       ReturnItemCollectionMetrics: constants.NONE,
     }
     dynamo.update(dbParamsContributions, updateCallback)
-  },
-  /**
-   * Get events from the Events table that will need to have the contributor attached to them.
-   * @param id The product id
-   * @param origin Who/what triggered this update
-   * @param roleInfo Who was the photographer or creator for this product
-   * @param baseline The eventId that first registered the contributor, so credit is only applied subsequently.
-   * @param complete The callback to inform of completion, with optional error parameter.
-   */
-  getEventsThenCredit: (id, baseline, origin, roleInfo, complete) => {
-    const params = {
-      TableName: constants.TABLE_EVENTS_NAME,
-      ProjectionExpression: '#e',
-      KeyConditionExpression: '#i = :i AND #e > :e',
-      ConsistentRead: true,
-      ExpressionAttributeNames: {
-        '#i': 'productId',
-        '#e': 'eventId',
-      },
-      ExpressionAttributeValues: {
-        ':i': id,
-        ':e': baseline,
-      },
-    }
-
-    dynamo.query(params, (err, data) => {
-      if (err) {
-        complete(`${constants.METHOD_GET_EVENTS_THEN_CREDIT} - errors updating DynamoDb: ${err}`)
-      } else if (!data || !data.Items || data.Items.length === 0) {
-        console.log(`Found no events already logged for ${id}, but occurring after the registration event ${baseline}.`) // TODO remove
-        complete()
-      } else {
-        console.log('Found later events that were already logged, needing contributor added ', data.Items) // TODO remove
-        impl.creditContributions(id, data.Items.map(item => item.eventId), origin, roleInfo, complete)
-      }
-    })
-  },
-  /**
-   * Assign credit to a product-event pair in the Events table, either because of a purchase event or to true up any
-   * events we may have seen prior to the registration due to the batch being out of order.  If there is no one to
-   * credit, just log the product-event id.
-   * @param id The product id
-   * @param origin Who/what triggered this update
-   * @param roleInfo Who was the photographer or creator for this product
-   * @param eventIds Array of event ids for that product needing credit entered into the Events table. Note that there
-   * is either both a photographer and a creator, in which case eventIds is length 1, or it is a bunch of events for a
-   * single registration of either a creator or a photographer (but not both).  Expect one or small size generally.
-   * Credit should only be assigned if the contributor registered prior to the purchase event, as reflected by the fact
-   * the event Id is further along in the sequence for that product than the registration event; the eventIds should
-   * reflect this.
-   * @param complete The callback to inform of completion, with optional error parameter.
-   */
-  creditContributions: (id, eventIds, origin, roleInfo, complete) => {
-    const updated = Date.now()
-
-    // Record contributor info for specific event.
-    const expression = [
-      'set',
-      '#c=if_not_exists(#c,:c),',
-      '#cb=if_not_exists(#cb,:cb),',
-      '#u=:u,',
-      '#ub=:ub',
-    ]
-    const attNames = {
-      '#c': 'created',
-      '#cb': 'createdBy',
-      '#u': 'updated',
-      '#ub': 'updatedBy',
-    }
-    const attValues = {
-      ':c': updated,
-      ':cb': origin,
-      ':u': updated,
-      ':ub': origin,
-    }
-
-    if (roleInfo) {
-      if (roleInfo.creator) {
-        expression.push(', #cr=:cr')
-        attNames['#cr'] = 'creator'
-        attValues[':cr'] = roleInfo.creator
-      }
-      if (roleInfo.photographer) {
-        expression.push(', #ph=:ph')
-        attNames['#ph'] = 'photographer'
-        attValues[':ph'] = roleInfo.photographer
-      }
-    } else {
-      console.log(`${constants.METHOD_CREDIT_CONTRIBUTIONS} No contributors passed, so just logging event.`) // TODO remove
-    }
-
-    let successes = 0
-    const groupDynamoCallback = (err) => {
-      if (err) {
-        complete(`${constants.METHOD_CREDIT_CONTRIBUTIONS} - errors updating DynamoDb: ${err}`)
-      } else {
-        successes += 1
-      }
-      if (successes === eventIds.length) {
-        console.log(`${constants.MODULE} ${constants.METHOD_CREDIT_CONTRIBUTIONS} - all ${eventIds.length} events updated successfully for ${id}.`)
-        impl.updateScoresTable(origin, roleInfo, complete)
-      }
-    }
-    for (let i = 0; i < eventIds.length; i++) {
-      const dbParamsEvents = {
-        TableName: constants.TABLE_EVENTS_NAME,
-        Key: {
-          productId: id,
-          eventId: eventIds[i],
-        },
-        UpdateExpression: expression.join(' '),
-        ExpressionAttributeNames: attNames,
-        ExpressionAttributeValues: attValues,
-        ReturnValues: constants.NONE,
-        ReturnConsumedCapacity: constants.NONE,
-        ReturnItemCollectionMetrics: constants.NONE,
-      }
-      dynamo.update(dbParamsEvents, groupDynamoCallback)
-    }
-  },
-  /**
-   * Update scores table on whatever contributor(s) were just affected.
-   * @param data Who to update that was affected by last update of creator and/or photographer.
-   * @param origin Who/what generated the activity leading to this update
-   * @param complete The callback to inform of completion, with optional error parameter.
-   */
-  updateScoresTable: (origin, data, complete) => {
-    const updated = Date.now()
-
-    let priorErr
-    const updateCallback = (err) => {
-      if (priorErr === undefined) { // first update result
-        if (err) {
-          priorErr = err
-        } else {
-          priorErr = false
-        }
-      } else if (priorErr && err) { // second update result, if an error was previously received and we have a new one
-        complete(`${constants.METHOD_UPDATE_SCORES_TABLES} - errors updating DynamoDb: ${[priorErr, err]}`)
-      } else if (priorErr || err) {
-        complete(`${constants.METHOD_UPDATE_SCORES_TABLES} - error updating DynamoDb: ${priorErr || err}`)
-      } else { // second update result if error was not previously seen.
-        complete()
-      }
-    }
-    if (!data || (!data.creator && !data.photographer)) {
-      console.log('No contributor information on that update, so no effect on scores.')
-      // TODO could log this to an UNKNOWN contributor for both
-      complete()
-    } else {
-      const updateExp = [
-        'set',
-        '#u=:u,',
-        '#ub=:ub,',
-        '#sc=:sc',
-      ].join(' ')
-      const attNames = {
-        '#u': 'updated',
-        '#ub': 'updatedBy',
-        '#sc': 'score',
-      }
-      const attValues = {
-        ':u': updated,
-        ':ub': origin,
-      }
-      if (data.creator) {
-        const params = {
-          TableName: constants.TABLE_EVENTS_NAME,
-          IndexName: 'EventsByCreator',
-          ProjectionExpression: '#i, #e', // TODO remove after removing console.log
-          KeyConditionExpression: '#cr = :cr',
-          ExpressionAttributeNames: {
-            '#i': 'productId', // TODO remove after removing console.log
-            '#e': 'eventId', // TODO remove after removing console.log
-            '#cr': 'creator',
-          },
-          ExpressionAttributeValues: {
-            ':cr': data.creator,
-          },
-        }
-
-        dynamo.query(params, (err, response) => {
-          if (err) { // error from dynamo
-            updateCallback(`${constants.METHOD_UPDATE_SCORES_TABLES} - errors getting records from GSI Creator DynamoDb: ${err}`)
-          } else {
-            console.log('Found creator pairs ', response.Items) // TODO remove
-            const attValuesCreator = Object.assign({}, attValues)
-            attValuesCreator[':sc'] = response.Count
-            const dbParamsCreator = {
-              TableName: constants.TABLE_SCORES_NAME,
-              Key: {
-                userId: data.creator,
-                role: 'creator',
-              },
-              UpdateExpression: updateExp,
-              ExpressionAttributeNames: attNames,
-              ExpressionAttributeValues: attValuesCreator,
-              ReturnValues: constants.NONE,
-              ReturnConsumedCapacity: constants.NONE,
-              ReturnItemCollectionMetrics: constants.NONE,
-            }
-            dynamo.update(dbParamsCreator, updateCallback)
-          }
-        })
-      } else { // TODO could log this to an UNKNOWN contributor instead
-        updateCallback()
-      }
-      if (data.photographer) {
-        const params = {
-          TableName: constants.TABLE_EVENTS_NAME,
-          IndexName: 'EventsByPhotographer',
-          ProjectionExpression: '#i, #e', // TODO remove after removing console.log
-          KeyConditionExpression: '#ph = :ph',
-          ExpressionAttributeNames: {
-            '#i': 'productId', // TODO remove after removing console.log
-            '#e': 'eventId', // TODO remove after removing console.log
-            '#ph': 'photographer',
-          },
-          ExpressionAttributeValues: {
-            ':ph': data.photographer,
-          },
-        }
-
-        dynamo.query(params, (err, response) => { // NB can't get strongly consistent read on GSI
-          if (err) { // error from dynamo
-            updateCallback(`${constants.METHOD_UPDATE_SCORES_TABLES} - errors getting records from GSI Photographer DynamoDb: ${err}`)
-          } else {
-            console.log('Found photographer pairs ', response.Items) // TODO remove
-            const attValuesPhotographer = Object.assign({}, attValues)
-            attValuesPhotographer[':sc'] = response.Count
-            const dbParamsPhotographer = {
-              TableName: constants.TABLE_SCORES_NAME,
-              Key: {
-                userId: data.photographer,
-                role: 'photographer',
-              },
-              UpdateExpression: updateExp,
-              ExpressionAttributeNames: attNames,
-              ExpressionAttributeValues: attValuesPhotographer,
-              ReturnValues: constants.NONE,
-              ReturnConsumedCapacity: constants.NONE,
-              ReturnItemCollectionMetrics: constants.NONE,
-            }
-            dynamo.update(dbParamsPhotographer, updateCallback)
-          }
-        })
-      } else { // TODO could log this to an UNKNOWN contributor instead
-        updateCallback()
-      }
-    }
   },
   /**
    * Log latest purchase of a given product id to the Events table.  Example event:
@@ -374,34 +144,308 @@ const impl = {
    * @param complete The callback to inform of completion, with optional error parameter.
    */
   updatePurchaseEvent: (event, complete) => {
+    const id = event.data.id
+    const eventId = event.eventId
+    const origin = event.origin
+    const updated = Date.now()
+
+    // We should know at least one contributor exists, because order guarantee says that the creation event had to have happened already, but sadly, it might get trimmed off.  To keep things simple, for products without a creation event we will attribute everything to UNKNOWN for the photographer, too (even if that makes it onto the stream).
+    const expression = [
+      'set',
+      '#c=if_not_exists(#c,:c),', // if_not_exists evaluates to the path (first argument) if the path exists in the item, otherwise it evaluates to the operand (second argument)
+      '#cb=if_not_exists(#cb,:cb),',
+      '#u=:u,',
+      '#ub=:ub,',
+      '#sc=if_not_exists(#sc,:zero) + :inc,', // Shouldn't need to check if this exist because order guarantee says this already will be there, but if the creation event got trimmed off, then we'll still fail trying to add 1 to a nonexistent value.  No creator will get the credit, just UNKNOWN.
+      '#sp=if_not_exists(#pe,if_not_exists(#sp,:zero)) + :inc,', // Only increment this if a creation event was logged and a photographer registers (which removes the #pe attribute) or if no creation event was logged in the first place, in which latter case it doesn't matter, because the number goes to UNKNOWN.
+      '#ro=if_not_exists(#ro,:unk),', // Note that if the creator got chopped off the front of the stream, we need to attribute the purchase to UNKNOWN.  If we were sure no manual creation event got lobbed on later, we could maybe just rely on emptiness, but we need this to be consistent with the idea (in registerContributor) that manual events are blocked from changing who gets credit (until we make this a feature).
+      '#ev=:ev',
+    ]
+    const attNames = {
+      '#c': 'created',
+      '#cb': 'createdBy',
+      '#u': 'updated',
+      '#ub': 'updatedBy',
+      '#sc': 'creatorScore', // These may be different for the two roles because some purchases may happen between the time that the creator and photographer registered
+      '#sp': 'photographerScore',
+      '#pe': 'photographerExists', // For the conditional on whether photographer exists yet
+      '#ev': 'lastEventId',
+      '#ro': 'creator',
+    }
+    const attValues = {
+      ':c': updated,
+      ':cb': origin,
+      ':u': updated,
+      ':ub': origin,
+      ':inc': 1,
+      ':zero': 0,
+      ':ev': eventId,
+      ':unk': constants.UNKNOWN,
+    }
+    const callback = (err) => {
+      if (err) {
+        if (err.code && err.code === 'ConditionalCheckFailedException') {
+          console.log(`${constants.METHOD_UPDATE_PURCHASE_EVENT} - event has already been processed or creation event for product ${id} occurred before stream horizon.  Skipping.`)
+          complete()
+        } else {
+          complete(`${constants.METHOD_UPDATE_PURCHASE_EVENT} - errors updating DynamoDb: ${err}`)
+        }
+      } else {
+        impl.updateScoresTable(origin, id, complete)
+      }
+    }
+
+    const dbParamsEvents = {
+      TableName: constants.TABLE_CONTRIBUTIONS_NAME,
+      Key: {
+        productId: id,
+      },
+      UpdateExpression: expression.join(' '),
+      ConditionExpression: 'attribute_not_exists(#ev) OR #ev < :ev',
+      ExpressionAttributeNames: attNames,
+      ExpressionAttributeValues: attValues,
+      ReturnValues: constants.NONE,
+      ReturnConsumedCapacity: constants.NONE,
+      ReturnItemCollectionMetrics: constants.NONE,
+    }
+    dynamo.update(dbParamsEvents, callback)
+  },
+  /**
+   * Update scores table on whatever contributor(s) were just affected.  This is also where photographers are updated to UNKNOWN in the cases where the creation event didn't make it onto the stream.
+   * @param id Which product id was affected by last purchase event.
+   * @param origin Who/what generated the activity leading to this update
+   * @param complete The callback to inform of completion, with optional error parameter.
+   */
+  updateScoresTable: (origin, id, complete) => {
+    const updated = Date.now()
+
+    let priorErr
+    const updateCallback = (err) => {
+      if (priorErr === undefined) { // first update result
+        if (err) {
+          priorErr = err
+        } else {
+          priorErr = false
+        }
+      } else if (priorErr && err) { // second update result, if an error was previously received and we have a new one
+        complete(`${constants.METHOD_UPDATE_SCORES_TABLES} - errors updating ${constants.TABLE_SCORES_NAME} DynamoDb: ${[priorErr, err]}`)
+      } else if (priorErr || err) {
+        complete(`${constants.METHOD_UPDATE_SCORES_TABLES} - error updating ${constants.TABLE_SCORES_NAME} DynamoDb: ${priorErr || err}`)
+      } else { // second update result if error was not previously seen.
+        complete()
+      }
+    }
+
     const dbParamsContributions = {
       Key: {
-        productId: event.data.id,
+        productId: id,
       },
       TableName: constants.TABLE_CONTRIBUTIONS_NAME,
       AttributesToGet: [
         'creator',
-        'creatorEventId',
         'photographer',
-        'photographerEventId',
       ],
-      ConsistentRead: true,
+      ConsistentRead: true, // Not really critical for scores, because not quite a permanent calculation (re-calculated on the fly, on each purchase event, so *may* get another chance)
       ReturnConsumedCapacity: constants.NONE,
     }
-    dynamo.get(dbParamsContributions, (err, data) => {
-      if (err) {
-        complete(`${constants.METHOD_UPDATE_PURCHASE_EVENT} - errors getting product ${event.data.id} from DynamoDb table ${constants.TABLE_CONTRIBUTIONS_NAME}: ${err}`)
+    dynamo.get(dbParamsContributions, (errBase, responseBase) => {
+      if (errBase) {
+        complete(`${constants.METHOD_UPDATE_SCORES_TABLES} - errors getting product ${id} from DynamoDb table ${constants.TABLE_CONTRIBUTIONS_NAME}: ${errBase}`)
       } else {
-        const roleInfo = {}
-        if (data && data.Item) {
-          if (data.Item.creator && data.Item.creatorEventId && event.eventId > data.Item.creatorEventId) {
-            roleInfo.creator = data.Item.creator
+        const data = responseBase.Item
+        if (!data || data.length === 0) {
+          complete(`${constants.METHOD_UPDATE_SCORES_TABLES} - unexpectedly could not find product ${id} from DynamoDb table ${constants.TABLE_CONTRIBUTIONS_NAME}`)
+        } else { // Sort out the creator
+          const updateExp = [
+            'set',
+            '#u=:u,',
+            '#ub=:ub,',
+            '#sc=:sc',
+          ].join(' ')
+          const attNames = {
+            '#u': 'updated',
+            '#ub': 'updatedBy',
+            '#sc': 'score',
           }
-          if (data.Item.photographer && data.Item.photographerEventId && event.eventId > data.Item.photographerEventId) {
-            roleInfo.photographer = data.Item.photographer
+          const attValues = {
+            ':u': updated,
+            ':ub': origin,
+          }
+
+          // Because this method is only called through the occurrence of a purchase event and that must be subsequent to a creator registration or set to UNKNOWN on the very next event for lack of a creator registration, we definitely have a creator field
+          const dbParamsCreator = {
+            TableName: constants.TABLE_CONTRIBUTIONS_NAME,
+            IndexName: 'ProductsByCreator',
+            ProjectionExpression: '#i, #s', // TODO remove id after removing console.log, only need the score really
+            KeyConditionExpression: '#ro = :ro',
+            ExpressionAttributeNames: {
+              '#i': 'productId', // TODO remove after removing console.log
+              '#s': 'creatorScore',
+              '#ro': 'creator',
+            },
+            ExpressionAttributeValues: {
+              ':ro': data.creator,
+            },
+          }
+
+          dynamo.query(dbParamsCreator, (err, response) => {
+            if (err) { // error from dynamo
+              updateCallback(`${constants.METHOD_UPDATE_SCORES_TABLES} - errors getting records from GSI Creator DynamoDb: ${err}`)
+            } else {
+              console.log('Found creator products ', response.Items) // TODO remove
+
+              const foundScores = response.Items.map(item => item.creatorScore) // There is always a score by this point
+              const attValuesCreator = Object.assign({}, attValues)
+              attValuesCreator[':sc'] = foundScores.length === 0 ? 0 : foundScores.reduce((acc, val) => acc + val)
+
+              const params = {
+                TableName: constants.TABLE_SCORES_NAME,
+                Key: {
+                  userId: data.creator,
+                  role: 'creator',
+                },
+                UpdateExpression: updateExp,
+                ExpressionAttributeNames: attNames,
+                ExpressionAttributeValues: attValuesCreator,
+                ReturnValues: constants.NONE,
+                ReturnConsumedCapacity: constants.NONE,
+                ReturnItemCollectionMetrics: constants.NONE,
+              }
+              dynamo.update(params, updateCallback)
+            }
+          })
+
+          // Sort out the photographer
+          if (data.creator === constants.UNKNOWN) {
+            // Set photographer to UNKNOWN fo this product, then update photographer score.
+            const phExp = [
+              'set',
+              '#c=if_not_exists(#c,:c),', // very strange of this didn't already exist
+              '#cb=if_not_exists(#cb,:cb),',
+              '#u=:u,',
+              '#ub=:ub,',
+              '#ro=:unk',
+            ]
+            const phAttNames = {
+              '#c': 'created',
+              '#cb': 'createdBy',
+              '#u': 'updated',
+              '#ub': 'updatedBy',
+              '#ro': 'photographer',
+            }
+            const phAttValues = {
+              ':c': updated,
+              ':cb': origin,
+              ':u': updated,
+              ':ub': origin,
+              ':unk': constants.UNKNOWN,
+            }
+            const phCallback = (errPhUnk) => {
+              if (errPhUnk) {
+                updateCallback(`${constants.METHOD_UPDATE_PURCHASE_EVENT} - errors updating Contributions DynamoDb so photographer is set to UNKNOWN: ${errPhUnk}`)
+              } else {
+                const dbParamsUnk = {
+                  TableName: constants.TABLE_CONTRIBUTIONS_NAME,
+                  IndexName: 'ProductsByPhotographer',
+                  ProjectionExpression: '#i, #s', // TODO remove id after removing console.log, only need the score really
+                  KeyConditionExpression: '#ro = :ro',
+                  ExpressionAttributeNames: {
+                    '#i': 'productId', // TODO remove after removing console.log
+                    '#s': 'photographerScore',
+                    '#ro': 'photographer',
+                  },
+                  ExpressionAttributeValues: {
+                    ':ro': constants.UNKNOWN, // No strong consistency for GSI, so may not pick up (this time around) the overwrite of photographer to UNKNOWN (*may* get another chance, on another purchase event)
+                  },
+                }
+
+                dynamo.query(dbParamsUnk, (err, response) => {
+                  if (err) { // error from dynamo
+                    updateCallback(`${constants.METHOD_UPDATE_SCORES_TABLES} - errors getting records from GSI Photographer DynamoDb for the UNKNOWN photographer: ${err}`)
+                  } else {
+                    console.log('Found photographer products ', response.Items) // TODO remove
+
+                    const foundScores = response.Items.map(item => item.photographerScore) // There is always a score by this point
+                    const attValuesUnk = Object.assign({}, attValues)
+                    attValuesUnk[':sc'] = foundScores.length === 0 ? 0 : foundScores.reduce((acc, val) => acc + val)
+
+                    const params = {
+                      TableName: constants.TABLE_SCORES_NAME,
+                      Key: {
+                        userId: constants.UNKNOWN,
+                        role: 'photographer',
+                      },
+                      UpdateExpression: updateExp,
+                      ExpressionAttributeNames: attNames,
+                      ExpressionAttributeValues: attValuesUnk,
+                      ReturnValues: constants.NONE,
+                      ReturnConsumedCapacity: constants.NONE,
+                      ReturnItemCollectionMetrics: constants.NONE,
+                    }
+                    dynamo.update(params, updateCallback)
+                  }
+                })
+              }
+            }
+            const dbParamsPh = {
+              TableName: constants.TABLE_CONTRIBUTIONS_NAME,
+              Key: {
+                productId: id,
+              },
+              UpdateExpression: phExp.join(' '),
+              ExpressionAttributeNames: phAttNames,
+              ExpressionAttributeValues: phAttValues,
+              ReturnValues: constants.NONE,
+              ReturnConsumedCapacity: constants.NONE,
+              ReturnItemCollectionMetrics: constants.NONE,
+            }
+            dynamo.update(dbParamsPh, phCallback)
+          } else if (data.photographer) { // This is the plain-vanilla photographer case.
+            const dbParamsPhotographer = {
+              TableName: constants.TABLE_CONTRIBUTIONS_NAME,
+              IndexName: 'ProductsByPhotographer',
+              ProjectionExpression: '#i, #s', // TODO remove id after removing console.log, only need the score really
+              KeyConditionExpression: '#ro = :ro',
+              ExpressionAttributeNames: {
+                '#i': 'productId', // TODO remove after removing console.log
+                '#s': 'photographerScore',
+                '#ro': 'photographer',
+              },
+              ExpressionAttributeValues: {
+                ':ro': data.photographer,
+              },
+            }
+
+            dynamo.query(dbParamsPhotographer, (err, response) => {
+              if (err) { // error from dynamo
+                updateCallback(`${constants.METHOD_UPDATE_SCORES_TABLES} - errors getting records from GSI Photographer DynamoDb: ${err}`)
+              } else {
+                console.log('Found products ', response.Items) // TODO remove
+
+                const foundScores = response.Items.map(item => item.photographerScore) // There is always a score by this point
+                const attValuesPhotographer = Object.assign({}, attValues)
+                attValuesPhotographer[':sc'] = foundScores.length === 0 ? 0 : foundScores.reduce((acc, val) => acc + val)
+
+                const params = {
+                  TableName: constants.TABLE_SCORES_NAME,
+                  Key: {
+                    userId: data.photographer,
+                    role: 'photographer',
+                  },
+                  UpdateExpression: updateExp,
+                  ExpressionAttributeNames: attNames,
+                  ExpressionAttributeValues: attValuesPhotographer,
+                  ReturnValues: constants.NONE,
+                  ReturnConsumedCapacity: constants.NONE,
+                  ReturnItemCollectionMetrics: constants.NONE,
+                }
+                dynamo.update(params, updateCallback)
+              }
+            })
+          } else { // Free pass, nothing to do for photographer
+            updateCallback()
           }
         }
-        impl.creditContributions(event.data.id, [event.eventId], event.origin, roleInfo, complete)
       }
     })
   },
