+++ Lesson6_CreateViewWithEventConsumerUsingParallelEventDelivery/winner-view/product-create-schema.json
