+++ Lesson6_CreateViewWithEventConsumerUsingParallelEventDelivery/winner-view/product-image-schema.json
