+++ Lesson6_CreateViewWithEventConsumerUsingParallelEventDelivery/winner-view/product-purchase-schema.json
