+++ Lesson6_CreateViewWithEventConsumerUsingParallelEventDelivery/winner-view/retail-stream-schema-egress.json
